requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Amadeus API Client
amadeus==8.1.0

# Anthropic AI
anthropic>=0.75.0

# HTTP Clients
httpx==0.26.0
requests==2.31.0

# Date/Time Handling
python-dateutil==2.8.2

# Validation & Schemas
pydantic==2.5.3
pydantic-settings==2.1.0

# Messaging (Optional)
# twilio==8.11.1
# sendgrid==6.11.0

# Development
pytest==7.4.3
pytest-asyncio==0.23.3
